add visitfee method

+++ src/acedemic.h
@@ -9,6 +9,6 @@ class Acedemic: public Property {
     Acedemic(string name);
     void improve();
     void unimprove();
-    int getVisitFee() const override;
+    size_t getVisitFee() const override;
     SquareInfo getInfo() const override;
 };
